% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse_tub.R
\name{parse_tub}
\alias{parse_tub}
\title{Function: Parse the tub files generated by gettub6.trx
Input: tub file path
Output: A data frame that contains tub data by head / zone
Algorithm: Read tub file -> Locate lines of head number -> Parse head/zone/DAC/Tub by line offset}
\usage{
parse_tub(tubFile)
}
\description{
Function: Parse the tub files generated by gettub6.trx
Input: tub file path
Output: A data frame that contains tub data by head / zone
Algorithm: Read tub file -> Locate lines of head number -> Parse head/zone/DAC/Tub by line offset
}

